Updating Religion

+++ catscripts/Lifestyle/Religion/Religion.sh
@@ -1,39 +1,40 @@
 #!/bin/bash
 
-if [ "$DEBUG" == "yes" ];
-then
-  printf "Starting Religion\n" 
-fi
-
 
 RELIGIONDIR="./catscripts/Lifestyle/Religion"
-
-. $RELIGIONDIR/Buddhism/Buddhism.sh
-. $RELIGIONDIR/Christianity/Christianity.sh #KEYWORDS_CHRISTIANITY_ALL
-. $RELIGIONDIR/Hinduism/Hinduism.sh
-. $RELIGIONDIR/Islam/Islam.sh
-. $RELIGIONDIR/Jainism/Jainism.sh
-. $RELIGIONDIR/Judaism/Judaism.sh #KEYWORDS_JUDAISM_ALL
-. $RELIGIONDIR/Scientology/Scientology.sh #KEYWORDS_SCIENTOLOGY_ALL
-. $RELIGIONDIR/Sikhism/Sikhism.sh
-. $RELIGIONDIR/The_Bible/TheBible.sh
-. $RELIGIONDIR/Wicca/Wicca.sh #KEYWORDS_WICCA
+KEYWORDS_RELIGION="religion|religious"
 
 
-egrep -i 'religion|religious' newpages.txt | egrep -iv "$KEYWORDS_CHRISTIANITY_ALL|$KEYWORDS_JUDAISM_ALL|Shinto|Sikh|god|buddhism|buddhist|hindu|jain(|s|ism)\b|islam|Muslim|Prophet (Muhammed|Muhammad)|$KEYWORDS_SCIENTOLOGY_ALL|$KEYWORDS_WICCA" >> Religion.txt
-
-RELIGION=$(stat --print=%s Religion.txt)
-
-if [ $RELIGION -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="Religion.txt"
-  export CATNAME="Religion"
-  $CATEGORIZE
-fi
 
-rm Religion.txt
+  debug_start "Religion"
 
-if [ "$DEBUG" == "yes" ];
-then
-  printf "Finishing religion\n" 
+  . $RELIGIONDIR/Buddhism/Buddhism.sh
+  . $RELIGIONDIR/Christianity/Christianity.sh #KEYWORDS_CHRISTIANITY_ALL
+  . $RELIGIONDIR/Hinduism/Hinduism.sh
+  . $RELIGIONDIR/Islam/Islam.sh
+  . $RELIGIONDIR/Jainism/Jainism.sh
+  . $RELIGIONDIR/Judaism/Judaism.sh #KEYWORDS_JUDAISM_ALL
+  . $RELIGIONDIR/Scientology/Scientology.sh #KEYWORDS_SCIENTOLOGY_ALL
+  . $RELIGIONDIR/Sikhism/Sikhism.sh
+  . $RELIGIONDIR/The_Bible/TheBible.sh
+  . $RELIGIONDIR/Wicca/Wicca.sh #KEYWORDS_WICCA
+
+  KEYWORDS_RELIGION_EXCLUDE="$KEYWORDS_CHRISTIANITY_ALL|$KEYWORDS_JUDAISM_ALL|Shinto|Sikh|god|buddhism|buddhist|hindu|jain(|s|ism)\b|islam|Muslim|Prophet (Muhammed|Muhammad)|$KEYWORDS_SCIENTOLOGY_ALL|$KEYWORDS_WICCA"
+
+  RELIGION=$(egrep -i "$KEYWORDS_RELIGION" newpages.txt | egrep -iv "$KEYWORDS_RELIGION_EXCLUDE")
+
+  if [ "$RELIGION" != "" ];
+  then
+    printf "%s" "$RELIGION" > Religion.txt
+    export CATFILE="Religion.txt"
+    export CATNAME="Religion"
+    $CATEGORIZE
+    rm Religion.txt
+    unset RELIGION
+  fi
+
+  debug_end "Religion"
+
 fi
